check repl conn, move to validate.go
